specific exception messages when the model is invalid for gust operations
tolerate dependent entities with no clear fk to update after saving the principle

+++ GustEfcConsumer/Model/PostVote.cs
@@ -19,7 +19,7 @@ namespace GustEfcConsumer.Model
         [Key]
         public int Id { get; set; }
                 
-        public long PostId { get; set; }
+        public int PostId { get; set; }
 
         public bool Positive { get; set; }
